Move static methods into helper class.
